requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.2
django-filter==24.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
mysqlclient==2.2.2
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pipenv==2023.12.1
platformdirs==4.1.0
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
retry-extended @ git+https://github.com/strollby/retry-extended@9b401b3dd1fce5c3e084ee130a8cc951a6d27fb9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0